Modify upgrade task to receive library option
