+++ configs/upgrade/portfolio/Calm.json
@@ -41,9 +41,9 @@
         "0xE3Ffc583dC176575eEA7FD9dF2A7c65F7E23f4C3",
         1,
         [8388607, 8388608, 8388609],
-        [100, 800, 100]
-      ],
-      0
+        [100, 800, 100],
+        0
+      ]
     ],
     "libraries": [
       {
