+++ test/strategies/traderjoe/TraderJoe.test.ts
@@ -320,7 +320,7 @@ function testTraderJoeInitialize() {
       await expect(
         upgrades.deployProxy(
           this.Strategy,
-          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [0, 1, 2], [300, 700]]],
+          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [0, 1, 2], [300, 700], 0]],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
         )
       ).to.be.reverted
@@ -330,7 +330,7 @@ function testTraderJoeInitialize() {
       await expect(
         upgrades.deployProxy(
           this.Strategy,
-          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [], [1000]]],
+          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [], [1000], 0]],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
         )
       ).to.be.reverted
@@ -351,6 +351,7 @@ function testTraderJoeInitialize() {
                 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51,
               ],
               [1000],
+              0,
             ],
           ],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
@@ -370,6 +371,7 @@ function testTraderJoeInitialize() {
               1,
               [16777216], // 2^24.
               [1000],
+              0,
             ],
           ],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
@@ -381,7 +383,7 @@ function testTraderJoeInitialize() {
       await expect(
         upgrades.deployProxy(
           this.Strategy,
-          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [8388608], []]],
+          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [8388608], [], 0]],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
         )
       ).to.be.reverted
@@ -402,6 +404,7 @@ function testTraderJoeInitialize() {
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27,
                 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51,
               ],
+              0,
             ],
           ],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
@@ -421,6 +424,7 @@ function testTraderJoeInitialize() {
               1,
               [8388607, 8388608, 8388609],
               [0, 700, 299], // It's sum should be 1e3.
+              0,
             ],
           ],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
@@ -432,7 +436,7 @@ function testTraderJoeInitialize() {
       await expect(
         upgrades.deployProxy(
           this.Strategy,
-          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [8388607, 8388608, 8388609], [0, 700, 301]]],
+          [this.strategyArgs, [TraderJoe.lbPair, TraderJoe.lbRouter, 1, [8388607, 8388608, 8388609], [0, 700, 301], 0]],
           { kind: "uups", unsafeAllow: ["external-library-linking"] }
         )
       ).to.be.reverted
